requirements.txt
altgraph==0.17.5
attrs==26.1.0
babel==2.18.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
colorlog==6.10.1
csvw==3.7.0
customtkinter==5.2.2
darkdetect==0.8.0
dlinfo==2.0.0
espeakng-loader==0.2.4
flatbuffers==25.12.19
idna==3.11
isodate==0.7.2
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro-onnx==0.4.7
lameenc==1.8.2
language-tags==1.2.0
mpmath==1.3.0
mutagen==1.47.0
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.0
pedalboard==0.9.22
pefile==2024.8.26
phonemizer-fork==3.3.1
pillow==12.2.0
protobuf==7.34.1
pycparser==3.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
rfc3986==1.5.0
rpds-py==0.30.0
scipy==1.17.1
segments==2.4.0
setuptools==82.0.1
six==1.17.0
sounddevice==0.5.5
soundfile==0.13.1
sympy==1.14.0
termcolor==3.3.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
